go: Plumb Error Recovery Foundations
This change first changes the compiler to collect one `recoveryEntry`
for each production that maps to an error label. e.g.:

  G <- "{" [0-9]* "}"^R
  R <-

At the end of the compilation, the table `recovery` will point the
label `R` to the production `R`.

The encoder will translate the `ILabel` within the `recoveryEntry`
into an actual code position and put the map from label/production IDs
to production addresses.

On the VM side, the `opThrow` operation will now check if the label ID
it received has any corresponding entries in the recovery map.  If
there is, control flow will be directed to the recovery production.
In this path, `opThrow` works very similarly to `opCall`, by pushing a
new frame onto the call stack and then pointing the `.pc` at the
address of the recovery expression.

The mechanism for collecting the errors isn't really finished.
Collecting terminals inserted within sequences automatically has to be
reworked as it has a little bug.