go: Rework Intermediary Captures
The semantics of `opCapEnd` (mostly implemented within `.newNode`)
were simplified a whole lot, and the logic for capturing terminals
mixed between non-terminals was lifted off the VM into the grammar
compiler.  Now, the `AddCaptures` step of the compiler is the step
that introduces unnamed captures around terminals that appear mixed
within rules that also contain non-terminals.  e.g.:

  Ordinal <- Digit ("st" / "nd" / "rd" / "th")
  Digit   <- [0-9]+

The grammar will will be augumented to

  Ordinal <- Digit (Cap("st") / Cap("nd") / Cap("rd") / Cap("th"))
  Digit   <- [0-9]+

Notice that the named captures around the productions were elided in
the example above for brevity.

One other interesting little change is that now `LabeledNode` within
the Grammar AST can't be considered Syntactic anymore because it might
behave as a call upon an exception with a recovery expression
associated with it.