go: InjectWhitespaces: traverse through CaptureNode as well
The grammar traversal `InjectWhitespaces` was written before the
traversal `AddCaptures`.  Because of that, a little bug slipped
through the cracks: If we add captures *before* injecting
whitespaces (for avoiding capturing spaces), whitespace handling
doesn't really do anything because the expression of every single
definition because the space injection step wasn't traversing through
`CaptureNode`.