go: `goeval` A new codegen for the Go language
Just like the existing `gen_go`, this new generator outputs everything
needed to parse an input string into a single file.  In fact, it
generates a parser with almost exactly the same API as the existing
mechanism.  The exception is the method `Parser.SetCaptureSpaces`,
which is currently not supported by the evaluator.

This change also uncommented the Grammar AST enrichment steps in the
`vm_test.go` file to do exactly what the `cmd/langlang/main.go` file
is doing.  At some point all that stuff is going to be moved into a
nice API that's going to hide that away from callers.

While outputting go code using `literalSanitizer`, it ended up failing
with the replacement of single quotes because apparently go doesn't
have an escape sequence for them.