go: Mechanism for suppressing captures during runtime
The steps for injecting space handling rule calls and for wrapping
terminal and non-terminal nodes around captures is fully parameterized
while generating the bytecode of a grammar.  However, it wasn't
possible to selectively disable capturing some of these nodes.

With this change, it is now possible to provide a slice of rule names
that will have its values discarded instead of captured.  Although the
mechanism is generic, it is currently exclusively dedicated to
suppressing spaces.