grammars[langlang]: uterly simplify Char definition and add unicode
Defining chars is simpler now, as there are just two types of escaping
sequences.  :

  1. Regular `n`, `r`, `t`, `'`, `"`, `-`, `\\`, and '[', ']'.
  2. Unicode \u HEX HEX HEX HEX.

One small bug-fix snook in is that Comment now doesn't require an EOL
at the end, although it will optionally try to consume it.  That fixes
the use of comments on inputs with an EOF at the end of a comment
instead of an EOL followed by an EOF.

Another small bug-fix is that we now ensure that the slash for the
`Expression` rule doesn't ever collide with the double slashes for
comments in both langlang.peg and peg.peg, since they both follow the
same convention for comments.