go[vm]: New Input abstraction
I didn't find a way of relying on the go interfaces to provide the
methods `PeekRune` and `Advance` necessary to implementing `opSet`.
So, the entire `Input` interface had to be reworked to fit in these
two new methods.

Also, a method `ReadString` was added, so we avoid an extra copy that
a `ReadAt` + `string` casting would have incurred.

Besides the change in the interface, this change also includes a
concrete implementation of this interface: MemInput.