go[*]: Make charsets actually fit into the design
This is a mix up of a lot of changes, but the most important change is
that the charset optimization won't apply for multi-byte characters
anymore.  If the grammar expects a multi-byte string either as a
literal or within a range, that `LiteralNode` or `RangeNode` won't be
translated into a `CharsetNode.` If a `ClassNode` combines multi-byte
and single-byte chars, it also does not translate to a `CharsetNode.`
Despite that, it still shows pretty good results on the benchmarks we
have.  That allowed all the generalization of multiple sizes for
charsets to be removed, and it is now much simpler.

There is also a new pattern that the charset translation catches:

   Seq(Not(Lit(l0) Any))

It gets translated into the complement of the charset that initially
includes `l0`. e.g.:

   Charset(l0).Complement()

Another big change is how we update the `virtualMachine.expected` set.
There is now a top-down flag that completely disables the populating
of the `expected` set.  The flag is exposed as the `-show-fails`
command line attribute.  This feature might help users find out what's
wrong with their input when grammars are straightforward enough, but
it comes at a high performance cost, so having the option to fully
disable it is paramount for stuff like JSON/XML/etc.  Also, when
enabled, the `expected` expected set of each `charset` is pre-computed
in compile-time.

One other change that made sense to be done now is the change in the
`Spacing` rule.  It is so hot that it's worth using the fastest
pattern:

  `Spacing <- [ \t\r\n]*`