go[compiler]: Capture handler fixes and improvements
First improvement, it doesn't traverse into predicates anymore.  There
is no meaning in capturing anything within a predicate because it
doesn't move the cursor anyway.

The second improvement was checking if a node is already a capture
before wrapping into a capture again.

And finally, a small improvement trying to save on how many captures
runtime has to generate, by applying the following rewrite:

   (!x Cap(y))* -> Cap((!x y)*)

This will generate way less capture frames on the stack.  The
`Comment` rule in `langlang.peg` is an example of where this is
useful:

   Comment <- '//' (!EOL .)* EOL

Without this rewrite, that expression will be rewritten as:

   Comment <- Cap('//') (!EOL Cap(.))* EOL

And will produce a new node for every `.` matched.  Whereas with this
rewrite, it becomes this instead:

   Comment <- Cap('//') Cap((!EOL .)*) EOL

Which generates a single capture frame for all the characters within
the comment.