go[BUG-FIX]: Disable optimizations on main
I've been scratching my head trying to understand what broke capturing
in a few places.  I saw it happen with the `String` rule of the JSON
grammar, and when debugging, I found out that the only difference
between a commit that worked to a commit that didn't was the fact that
this new `main.go` was hardcoding compiler optimizations to `1`,
instead of the old default of `0`.

I haven't fully investigated what's wrong there (I suspect that
fail_twice might be discarding an important capture frame!) but think
this should go back to working on the `main` branch, thus this commit.