go[BUG-FIX]: PartialCommit must update capture count
While rearranging the command line arguments in `main.go`, I
accidentally changed the default for the `CompilerConfig.Optimize`
value, and it broke capturing in a bit of a weird way by making a few
values disappear.

Upon further inspection, I noticed that the issue was with the
`PartialCommit` optimization for `ZeroOrMore` that was missing the
update of how many captures it's seen so far.  That would have
normally happened through other codepath, where the backtracking frame
created by `Choice` will keep the correct counter.

With that, I was able to now intentionally switch the default value of
`CompilerConfig.Optimize` to `1` in `main.go`.