go[gen]: Sort code generated off of maps
The Go code generator emits some code based off of maps.  More
specifically, it emits the recovery express table and a separate
method for each production in the grammar, but since we were
enumerating these productions off of a map, their order wasn't
guaranteed across runs, which is very annoying because it causes
changes in the parsers that had no changes in their respective
grammars.