go[vm]: fix perf issue in captures by eliminating stack traversal
This commit replaces the backward stack traversal that looked for the
top-most `frameType_Capture` in the stack with an O(1) search on the
stack for *every* time we called `findCaptureFrame()`, which we did
quite a few times :face-palm: it became a very hot function in the
profiler.  The fix is to propagate values upward in the stack through
all the types of stack frames.  This still collects captures *while*
parsing.

NOTE: There is still room to improve perf here by maybe following what
LPEG does and generate just a log that can be used to build an actual
tree afterwards.