go[benchmarks]: Add benchmark output for v0.0.8-84-g46165de
Differences from the previous run:

 * This run is using a new directory structure
 * It has two more tests now: charsets, json
 * Runs are split up between cap and no cap
 * Construction of parser was taken off tight loop

Also, I am almost positive that the benchmarks for `c14539f` must have
been generated with the wrong binary, because I can't reproduce it
with the same results (although there have also been a few changes to
a few grammars itself).  For now, we're going to leave it the way it
is, but this commit adds the results that seem to be accurate and
reproducible!