go[compvm]: Rename capOnce to capTerm and create capNonTerm
There is no need to create a capture stack frame if the size of the
expression of a nonterminal is known at compile time.  This doesn't
make a huge difference with any grammars we have in our existing
benchmarks, but grammars with lots of NonTerminals that are totally
syntactic without any zero-arity sub-expressions will benefit a lot
from this lil improvement.  Not to mention that this also stacks up
very well with inlining.