go[benchmarks]: Add a stripped version of the JSON grammar
This benchmark has a critical part of the grammar rewritten.  The
"String" rule doesn't do calls and does much more localized
backtracking, prioritizing the fastest alternative `!["\\] .` over
Escape and Unicode.  This fastest path compiles to the complement of
the set, which is pretty fast to match.  Because of escaping, we can't
get it to compile into a set, but that's much much better than all the
indirection with the nested rules.

This benchmark allows observing the performance of syntactical rules a
bit more deeply, and help building understanding of what's the impact
of nesting rules vs what's the perf of pure syntactical matching.