go[transform]: Fix space injection for syntactic rules with labels
There is some nuance to what being a sytactic expression means.  The
definition itself is clear: if the whole expression is made
exclusively of `Terminal` nodes, the expression is syntactic.  And the
container nodes like `ZeroOrMore`, `OneOrMoreNode`, etc usually just
defer the same question to their contained expression.

The problem is that `Labeled` nodes will be translated to an
`opChoice`, in which the first alternative is its contained
expression, but the second alternative is an `opThrow`.  Which will
act like a `Non-Terminal` if there is a recovery expression available
for that label.

That way, we kind of have to answer differently the question is a
labeled node syntactic if the question is being asked either by the
space injection or by the capture system.  The capture system must be
ready to receive another leaf node when it wraps a throw, but the
space injection system doesn't need to know about that nuance.