go[BUGFIX]: Enable show fails by default
I'm not quite sure of what's going on but the `TestImport` function
wasn't calling `p.SetShowFails(false)` but was still expecting a test
that fails parsing not to show the expected characters.  I updated the
test to reflect interaction expectations.

On the other side, I also fixed a small mistake where we were not
initializing the field `virtualMachine.expected` if the parser didn't
get constructed with `showFails=true`.