go[compiler]: Naive implementation of inlining
There are a few pieces to that:

 1. `callGraph` really just to detect recursion by visiting all the
   possible vertices (definitions), and finding all the edges between
   them (identifiers) then traversing them in a dfs fashion and
   stacking identifiers like the actual execution will do.

 2. `compiler.shouldInline` tells if a definition will be inlined or
    not.  The main definition is never inlined.  Then the number of
    instructions of the definition is checked against the new setting
    `compiler.inline.max_size`.  If these checks pass, the definition
    is marked to be inlined.

 3. `VisitIdentifierNode`: if `shouldInline` is true it follows the
    expression within the inlined definition instead of emitting the
    `ICall` opcode.

 4. `VisitDefinitionNode`: if `shouldInline` is true and if emitting
    inlined definitions has been disabled via the newly added setting
    `compiler.inline.emit.inlined`, then this function will not emit
    any code.

The function `getDefSize` might be a lil expesive because it
instantiates a full blown new compiler.  Although we do cache the size
of that definition requiring it to be compiled just once, so even if
the same production is referred to multiple times (like Spacing) its
size is only queried once before it's cached!

Also, for the callGraph's lookups for `Identifier` nodes, I ended up
adding the `Inspect` method since it is going to be useful for future
improvements that take writing traversals.