go[perf/vm]: hoist cursor and pc in hot path, clean up FFP update
Both the input cursor and the program counter are rather hot variables
within the VM, so they are likely more performatic being accessed as
local variables instead of struct attributes.

While making that update to cursor and pc, I also found a few places
where the FFP update could be a little cleaner.  Now, the FFP update
only really happens at the `fail` label!

And the update of the `expected` set is handled separately.