go[perf/vm]: avoid bounds check within tight loop
Seems like adding a condition to the for loop ensuring that pc is
always less than the length of the byte code array might have gotten
rid of some bounds checks while acessing the code array.  I didn't
investigate deep enough, but that seems to be the best explanation.

This improvement comes in handy right after adding the cap_commit type
instructions might have made the tight loop a teeny bit slower because
of the number of branches in the switch increasing, this is a good win
to offset that loss and even accrue a bit more of perf gains,
especially in the nocap variants!