grammars[langlang]: Fix parsing range with dash at the end
Consider the following grammar:

  P <- [+-]

The intent was clearly to parse either the plus or the minus operation
while parsing a "polarity" field.  But because the dash char (`-`) is
also used to denote a `Range`, the parser was getting confused and was
just considering whatever was in front of the dash to be the end of
the range.  So in this case it would match anything between the ascii
char 43(`+`) and 93(`]`).

This is a teeny corner case that only happens because one slight
change in the Char rule (replacement of the `!'\\' .` with a simple
`.`) but the fix I'm trying here is as simple as just negating the `]`
char before consuming a second `Char`.