go[config] BUG! Fix disabling capturing spaces
The builtin rule Spacing got inserted before the gnarly if/else we
had in api.go, so it was not doing much.  I just moved the check to
inside the capture handler itself.  It is mostly OK, but there is
the fact that the code is hardcoding the builtin rule name...