go[comp] BUG! Follow Spacing deps when attributing caps
This is a tiny little bug fix that follows the spacing deps when
we are trying to skip attributing capture nodes for each definition.

Although small, this is a rather important developer experience fix
as it will cause output tree shapes to break if a given parser is
recompiled, which may cause very subtle compatibility breakage, leading
to annoying bugs to fix on cst translators out there.

The bug was mainly exposed when one tried to extend the builtin Spacing
rule with something else to be skipped, like comments for example.

The langlang.peg grammar had this issue for example.