go[BREAKING]: Pre-compile error labels to make parser creation cheaper
There may be cases where one doesn't want to reuse the same parser
instance, like in concurrent cases.  And to serve those cases,
creating parsers should be as cheap as possible.

To contribute to creating new parser instances faster, this change
removes the mapping of error labels and error messages into IDs within
the internalized slice of strings within the VM from the hot path.

That is achieved by introducing the function `LabelMessagesForParser`
meant to be used to create this mapping and make it available for
`Parser.SetLabelMessages`, which could be called multiple times
without having to redo the mapping.  e.g.:

  m := LabelMessagesForParser(map[string]string{
    "label": "message",
    // ...
  })
  for range N {
    p := NewParser()
    p.SetInput([]byte(input))
    p.SetLabelMessages(m) // <--- this is now very cheap
  }

This also removes the allocation of a map in the parser's constructor.